Add component to display video feeds
The 'video-strip' component is a window that holds the 'video-list'.
This allows for the 'video-list' to be any size and for the
'video-strip' window to scroll through the 'video-list' content.
A 'video-service' component will supply the list of video-stream urls.
For now it simply returns an array of empty objects.

Next steps:
 * Video-list should dynamically resize based on the # of videos
 * Get real data for the videos