WIP scroll to bottom of chat window on message
Using the 'room-info' component as a model, this commit adds a custom
component for the room messages. By grabbing different pieces from the
'index.html' and copying them into the template, it was made to work the
same way as the old '#messages' container, but it's very broken in other
ways.

1) The methods in 'room-messages/index.js' aren't called when I would
expect, and I wanted to use them to hook into the event fired when a
message is received to scroll to the bottom. Look for my 'debugger'
statement to see what I mean.
2) Users aren't being taken out of the room when they close a tab, as
far as I can tell, or maybe they are joining the room twice. This could
be because the 'socket-con' and 'room-info' elements are duplicated in
the template for the 'room-messages' component.

I'll be looking forward to either pairing on fixing this or reviewing
the follow-up commit which does so.

Thanks!